remove AbstractRocksDBTimeOrderedWindowSegmentedBytesStoreTest.java and remove unused setup method
