+++ streams/src/test/java/org/apache/kafka/streams/state/internals/AbstractDualSchemaRocksDBSegmentedBytesStoreTest.java
@@ -92,11 +92,13 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends Segment> {
+public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest {
 
+    private static final String METRICS_SCOPE = "metrics-scope";
+    
     private final long windowSizeForTimeWindow = 500;
     private InternalMockProcessorContext context;
-    private AbstractDualSchemaRocksDBSegmentedBytesStore<S> bytesStore;
+    private AbstractDualSchemaRocksDBSegmentedBytesStore<KeyValueSegment> bytesStore;
     private File stateDir;
     private final Window[] windows = new Window[4];
     private Window nextSegmentWindow, startEdgeWindow, endEdgeWindow;
@@ -106,6 +108,16 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
     final long retention = 1000;
     final long segmentInterval = 60_000L;
     final String storeName = "bytes-store";
+
+    abstract boolean hasIndex();
+    abstract SchemaType schemaType();
+
+    enum SchemaType {
+        WindowSchemaWithIndex,
+        WindowSchemaWithoutIndex,
+        SessionSchemaWithIndex,
+        SessionSchemaWithoutIndex
+    }
     
     @BeforeEach
     public void before() {
@@ -156,13 +168,61 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
         bytesStore.close();
     }
 
-    abstract AbstractDualSchemaRocksDBSegmentedBytesStore<S> getBytesStore();
+    AbstractDualSchemaRocksDBSegmentedBytesStore<KeyValueSegment> getBytesStore() {
+        switch (schemaType()) {
+            case WindowSchemaWithIndex:
+            case WindowSchemaWithoutIndex:
+                return new RocksDBTimeOrderedWindowSegmentedBytesStore(
+                        storeName,
+                        METRICS_SCOPE,
+                        retention,
+                        segmentInterval,
+                        hasIndex()
+                );
+            case SessionSchemaWithIndex:
+            case SessionSchemaWithoutIndex:
+                return new RocksDBTimeOrderedSessionSegmentedBytesStore(
+                        storeName,
+                        METRICS_SCOPE,
+                        retention,
+                        segmentInterval,
+                        hasIndex()
+                );
+            default:
+                throw new IllegalStateException("Unknown SchemaType: " + schemaType());
+        }
+    };
 
-    abstract AbstractSegments<S> newSegments();
+    AbstractSegments<KeyValueSegment> newSegments() {
+        return new KeyValueSegments(storeName, METRICS_SCOPE, retention, segmentInterval);
+    }
 
-    abstract KeySchema getBaseSchema();
+    KeySchema getBaseSchema() {
+        switch (schemaType()) {
+            case WindowSchemaWithIndex:
+            case WindowSchemaWithoutIndex:
+                return new TimeFirstWindowKeySchema();
+            case SessionSchemaWithIndex:
+            case SessionSchemaWithoutIndex:
+                return new TimeFirstSessionKeySchema();
+            default:
+                throw new IllegalStateException("Unknown SchemaType: " + schemaType());
+        }
+    }
 
-    abstract KeySchema getIndexSchema();
+    KeySchema getIndexSchema() {
+        if (!hasIndex()) {
+            return null;
+        }
+        switch (schemaType()) {
+            case WindowSchemaWithIndex:
+                return new KeyFirstWindowKeySchema();
+            case SessionSchemaWithIndex:
+                return new KeyFirstSessionKeySchema();
+            default:
+                throw new IllegalStateException("Unknown SchemaType: " + schemaType());
+        }
+    }
 
     @Test
     public void shouldPutAndFetch() {
@@ -1057,7 +1117,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
     @Test
     public void shouldRollSegments() {
         // just to validate directories
-        final AbstractSegments<S> segments = newSegments();
+        final AbstractSegments<KeyValueSegment> segments = newSegments();
         final String key = "a";
 
         bytesStore.put(serializeKey(new Windowed<>(key, windows[0])), serializeValue(50));
@@ -1094,7 +1154,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
     @Test
     public void shouldGetAllSegments() {
         // just to validate directories
-        final AbstractSegments<S> segments = newSegments();
+        final AbstractSegments<KeyValueSegment> segments = newSegments();
         final String keyA = "a";
         final String keyB = "b";
 
@@ -1127,7 +1187,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
     @Test
     public void shouldGetAllBackwards() {
         // just to validate directories
-        final AbstractSegments<S> segments = newSegments();
+        final AbstractSegments<KeyValueSegment> segments = newSegments();
         final String keyA = "a";
         final String keyB = "b";
 
@@ -1159,7 +1219,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
     @Test
     public void shouldFetchAllSegments() {
         // just to validate directories
-        final AbstractSegments<S> segments = newSegments();
+        final AbstractSegments<KeyValueSegment> segments = newSegments();
         final String key = "a";
 
         bytesStore.put(serializeKey(new Windowed<>(key, windows[0])), serializeValue(50L));
@@ -1190,7 +1250,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
 
     @Test
     public void shouldLoadSegmentsWithOldStyleDateFormattedName() {
-        final AbstractSegments<S> segments = newSegments();
+        final AbstractSegments<KeyValueSegment> segments = newSegments();
         final String key = "a";
 
         bytesStore.put(serializeKey(new Windowed<>(key, windows[0])), serializeValue(50L));
@@ -1226,7 +1286,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
 
     @Test
     public void shouldLoadSegmentsWithOldStyleColonFormattedName() {
-        final AbstractSegments<S> segments = newSegments();
+        final AbstractSegments<KeyValueSegment> segments = newSegments();
         final String key = "a";
 
         bytesStore.put(serializeKey(new Windowed<>(key, windows[0])), serializeValue(50L));
@@ -1272,7 +1332,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
         final Collection<ConsumerRecord<byte[], byte[]>> records = new ArrayList<>();
         records.add(new ConsumerRecord<>("", 0, 0L, serializeKey(new Windowed<>(key, windows[0]), true).get(), serializeValue(50L)));
         records.add(new ConsumerRecord<>("", 0, 0L, serializeKey(new Windowed<>(key, windows[3]), true).get(), serializeValue(100L)));
-        final Map<S, WriteBatch> writeBatchMap = bytesStore.getWriteBatches(records);
+        final Map<KeyValueSegment, WriteBatch> writeBatchMap = bytesStore.getWriteBatches(records);
         assertEquals(2, writeBatchMap.size());
 
         final int expectedCount = getIndexSchema() == null ? 1 : 2;
@@ -1578,7 +1638,7 @@ public abstract class AbstractDualSchemaRocksDBSegmentedBytesStoreTest<S extends
     @Test
     public void shouldMeasureExpiredRecords() {
         final Properties streamsConfig = StreamsTestUtils.getStreamsConfig();
-        final AbstractDualSchemaRocksDBSegmentedBytesStore<S> bytesStore = getBytesStore();
+        final AbstractDualSchemaRocksDBSegmentedBytesStore<KeyValueSegment> bytesStore = getBytesStore();
         final InternalMockProcessorContext context = new InternalMockProcessorContext(
             TestUtils.tempDirectory(),
             new StreamsConfig(streamsConfig)
